+++ tools/tiff2ps.c
@@ -1246,8 +1246,8 @@ PS_Lvl2page(FILE* fd, TIFF* tif, uint32 w, uint32 h)
 	}
 	buf_data = (unsigned char *)_TIFFmalloc(chunk_size);
 	if (!buf_data) {
-		TIFFError(filename, "Can't alloc %u bytes for %s.",
-			chunk_size, tiled_image ? "tiles" : "strips");
+		TIFFError(filename, "Can't alloc %lu bytes for %s.",
+			(unsigned long) chunk_size, tiled_image ? "tiles" : "strips");
 		return(FALSE);
 	}
